// 完善业务逻辑

+++ LoveSports/Model/DataModel/PedometerModel.h
@@ -75,8 +75,8 @@ typedef void(^PedometerModelSyncEnd)(NSDate *date, BOOL success);
 @property (nonatomic, assign) NSInteger fastRunTime ;       // 当天快速跑步时间
 
 @property (nonatomic, assign) NSInteger targetStep;         // 目标步数
-@property (nonatomic, assign) NSInteger targetCalories;     // 目标卡路里
-@property (nonatomic, assign) NSInteger targetDistance;     // 目标距离
+@property (nonatomic, assign) CGFloat targetCalories;     // 目标卡路里
+@property (nonatomic, assign) CGFloat targetDistance;     // 目标距离
 @property (nonatomic, assign) NSInteger targetSleep;        // 目标睡眠
 
 @property (nonatomic, assign) NSInteger stepSize;           // 步距
